% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecccdownload.R
\name{ecccdownload}
\alias{ecccdownload}
\title{Download Environment Canada Historical Climate Data}
\usage{
ecccdownload(stationID, start_date, end_date, timeframe)
}
\arguments{
\item{stationID}{A numeric or character value representing the station ID.}

\item{start_date}{A string representing the start date in "YYYY-MM-DD" format.}

\item{end_date}{A string representing the end date in "YYYY-MM-DD" format.}

\item{timeframe}{A numeric value (1 = daily, 2 = hourly, 3 = monthly).}
}
\value{
A data frame containing the filtered climate data for the specified date range and timeframe.
}
\description{
This function downloads climate data from Environment and Climate Change Canada for a given station ID, date range, and timeframe.
}
