% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstations.R
\name{getstations}
\alias{getstations}
\title{Get Environment Canada Historical Climate Station Inventory}
\usage{
getstations(
  province = NULL,
  name = NULL,
  climateid = NULL,
  stationid = NULL,
  timeframe = NULL,
  year = NULL
)
}
\arguments{
\item{province}{A string representing the province. Default is NULL.}

\item{name}{A string to search for in the station names. Default is NULL.}

\item{climateid}{A numeric value representing the Climate ID. Default is NULL.}

\item{stationid}{A numeric value representing the Station ID. Default is NULL.}

\item{timeframe}{A string specifying the timeframe: "hourly", "daily", or "monthly". Default is NULL.}

\item{year}{A numeric value representing the year to filter by. Default is NULL.}
}
\value{
A data frame containing the filtered station inventory.
}
\description{
This function retrieves and filters the station inventory from Environment and Climate Change Canada based on the provided parameters.
}
